+++ CHANGELOG.md
@@ -14,7 +14,15 @@ Requires `libfranka` >= 0.8.0
     `F_T_NE`, only settable in Desk, and `NE_T_EE`, which can be set in `franka_ros` with `SetEEFrame`
     and defaults to the identity transformation.
   * Add `F_T_NE` and `NE_T_EE` to `franka_msgs/FrankaState`.
+  * _Franka Gazebo Integration_: Now you can simulate Panda robots in Gazebo including:
+    * gravity compensation
+    * non-realtime commands like `setEEFrame` or `setLoad`
+    * gripper simulation with the _same_ action interface as `franka_gripper`
+    * estimated inertias in the URDF
+    * no need to change existing ROS controllers
+    * only torque control supported in this version
   * Extract Model Library in abstract base class interface. This allows users to implement their own model.
+  * **BREAKING** Remove `panda_arm_hand.urdf.xacro`. Use `panda_arm.urdf.xacro hand:=true` instead.
 
 ## 0.7.1 - 2020-10-22